+++ app/Filament/Resources/UserResource/Pages/CreateUser.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\UserResource\Pages;
+
+use App\Filament\Pages\CreateAuditedRecord;
+use App\Filament\Resources\UserResource;
+
+class CreateUser extends CreateAuditedRecord
+{
+    protected static string $resource = UserResource::class;
+}